# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Privacy-focused password analysis and security education API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface PasswordPolicy {
  /**
     * @minimum 1
     * @maximum 128
     */
  minimumLength: number;
  rejectCommonPasswords: boolean;
  checkPersonalInfo: boolean;
  allowSpaces: boolean;
}

export interface PersonalContext {
  /** @maxLength 80 */
  firstName?: string;
  /** @maxLength 4 */
  birthYear?: string;
  /** @maxLength 120 */
  organization?: string;
}

export interface PasswordAnalysisInput {
  /**
     * @minLength 0
     * @maxLength 256
     */
  password: string;
  personalContext?: PersonalContext;
  policy?: PasswordPolicy;
}

export type FindingSeverity = typeof FindingSeverity[keyof typeof FindingSeverity];


export const FindingSeverity = {
  positive: 'positive',
  caution: 'caution',
  critical: 'critical',
} as const;

export interface Finding {
  type: string;
  severity: FindingSeverity;
  title: string;
  detail: string;
}

export interface Suggestion {
  title: string;
  detail: string;
}

export type PasswordMetricsPolicyChecksItem = {
  label: string;
  passed: boolean;
};

export interface PasswordMetrics {
  length: number;
  lengthBand: string;
  uniqueCharacterCount: number;
  uniqueCharacterRatio: number;
  characterPool: number;
  characterTypes: number;
  estimatedEntropy: number;
  patternCount: number;
  policyPass: boolean;
  policyChecks: PasswordMetricsPolicyChecksItem[];
}

export type PasswordAnalysisClassification = typeof PasswordAnalysisClassification[keyof typeof PasswordAnalysisClassification];


export const PasswordAnalysisClassification = {
  VERY_WEAK: 'VERY WEAK',
  WEAK: 'WEAK',
  MODERATE: 'MODERATE',
  STRONG: 'STRONG',
  VERY_STRONG: 'VERY STRONG',
} as const;

export interface PasswordAnalysis {
  /**
     * @minimum 0
     * @maximum 100
     */
  score: number;
  classification: PasswordAnalysisClassification;
  findings: Finding[];
  suggestions: Suggestion[];
  metrics: PasswordMetrics;
}

export type PasswordGenerationInputLength = typeof PasswordGenerationInputLength[keyof typeof PasswordGenerationInputLength];


export const PasswordGenerationInputLength = {
  NUMBER_16: 16,
  NUMBER_20: 20,
  NUMBER_24: 24,
} as const;

export interface PasswordGenerationInput {
  length: PasswordGenerationInputLength;
  uppercase: boolean;
  lowercase: boolean;
  numbers: boolean;
  symbols: boolean;
}

export interface GeneratedPassword {
  password: string;
  length: number;
  composition: string[];
}

export interface ClassificationCount {
  classification: string;
  count: number;
}

export interface WeaknessCount {
  type: string;
  label: string;
  count: number;
}

export interface DashboardStats {
  totalAnalyses: number;
  averageScore: number;
  averageLength: number;
  classificationCounts: ClassificationCount[];
  recentScores: number[];
}

export interface WeaknessAnalytics {
  weaknesses: WeaknessCount[];
}