# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Privacy-focused password analysis and security education API
 * OpenAPI spec version: 0.1.0
 */
import type { ClassificationCount } from './classificationCount';

export interface DashboardStats {
  totalAnalyses: number;
  averageScore: number;
  averageLength: number;
  classificationCounts: ClassificationCount[];
  recentScores: number[];
}