# lib/api-zod/src/generated/types/finding.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Privacy-focused password analysis and security education API
 * OpenAPI spec version: 0.1.0
 */
import type { FindingSeverity } from './findingSeverity';

export interface Finding {
  type: string;
  severity: FindingSeverity;
  title: string;
  detail: string;
}