# lib/api-zod/src/generated/types/generatedPassword.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Privacy-focused password analysis and security education API
 * OpenAPI spec version: 0.1.0
 */

export interface GeneratedPassword {
  password: string;
  length: number;
  composition: string[];
}