# lib/api-zod/src/generated/types/passwordAnalysis.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Privacy-focused password analysis and security education API
 * OpenAPI spec version: 0.1.0
 */
import type { Finding } from './finding';
import type { PasswordAnalysisClassification } from './passwordAnalysisClassification';
import type { PasswordMetrics } from './passwordMetrics';
import type { Suggestion } from './suggestion';

export interface PasswordAnalysis {
  /**
     * @minimum 0
     * @maximum 100
     */
  score: number;
  classification: PasswordAnalysisClassification;
  findings: Finding[];
  suggestions: Suggestion[];
  metrics: PasswordMetrics;
}