# lib/api-zod/src/generated/types/passwordAnalysisInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Privacy-focused password analysis and security education API
 * OpenAPI spec version: 0.1.0
 */
import type { PasswordPolicy } from './passwordPolicy';
import type { PersonalContext } from './personalContext';

export interface PasswordAnalysisInput {
  /**
     * @minLength 0
     * @maxLength 256
     */
  password: string;
  personalContext?: PersonalContext;
  policy?: PasswordPolicy;
}