# lib/api-zod/src/generated/types/passwordGenerationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Privacy-focused password analysis and security education API
 * OpenAPI spec version: 0.1.0
 */
import type { PasswordGenerationInputLength } from './passwordGenerationInputLength';

export interface PasswordGenerationInput {
  length: PasswordGenerationInputLength;
  uppercase: boolean;
  lowercase: boolean;
  numbers: boolean;
  symbols: boolean;
}