# lib/api-zod/src/generated/types/passwordMetrics.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Privacy-focused password analysis and security education API
 * OpenAPI spec version: 0.1.0
 */
import type { PasswordMetricsPolicyChecksItem } from './passwordMetricsPolicyChecksItem';

export interface PasswordMetrics {
  length: number;
  lengthBand: string;
  uniqueCharacterCount: number;
  uniqueCharacterRatio: number;
  characterPool: number;
  characterTypes: number;
  estimatedEntropy: number;
  patternCount: number;
  policyPass: boolean;
  policyChecks: PasswordMetricsPolicyChecksItem[];
}