# lib/api-zod/src/generated/types/passwordPolicy.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Privacy-focused password analysis and security education API
 * OpenAPI spec version: 0.1.0
 */

export interface PasswordPolicy {
  /**
     * @minimum 1
     * @maximum 128
     */
  minimumLength: number;
  rejectCommonPasswords: boolean;
  checkPersonalInfo: boolean;
  allowSpaces: boolean;
}