# lib/api-zod/src/generated/types/personalContext.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Privacy-focused password analysis and security education API
 * OpenAPI spec version: 0.1.0
 */

export interface PersonalContext {
  /** @maxLength 80 */
  firstName?: string;
  /** @maxLength 4 */
  birthYear?: string;
  /** @maxLength 120 */
  organization?: string;
}