# lib/api-zod/src/generated/types/weaknessAnalytics.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Privacy-focused password analysis and security education API
 * OpenAPI spec version: 0.1.0
 */
import type { WeaknessCount } from './weaknessCount';

export interface WeaknessAnalytics {
  weaknesses: WeaknessCount[];
}