# lib/api-zod/src/generated/types/weaknessCount.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Privacy-focused password analysis and security education API
 * OpenAPI spec version: 0.1.0
 */

export interface WeaknessCount {
  type: string;
  label: string;
  count: number;
}